feat: Move SourceNotInstalledModal to kmpuiviews and Update Dependencies
This commit moves the `SourceNotInstalledModal` to the `kmpuiviews` module for multiplatform use, removes the old component from `UIViews`, and updates related dependencies.

-   **SourceNotInstalledModal:**
    -   Moved `SourceNotInstalledModal` from `UIViews` to `kmpuiviews`.
    -   Removed the old `SourceNotInstalledModal` from `UIViews`.
-    -   Updated imports related to `SourceNotInstalledModal`.
-    - Updated the dependencies to the new `SourceNotInstalledModal`.
+   - Updated the imports related to `SourceNotInstalledModal`.
+   - Updated `FavoriteFragment`, `NotificationFragment`, and `HistoryFragment` to use the new `SourceNotInstalledModal`
+   - Added back handler to the kmp `SourceNotInstalledModal`
+-   **ComposableUtils:**
+   - Updated the `ComposableUtils` to use the new kmp `ComposableUtils`.
+- **Other**
+    - General cleanup.
+    - Removed unused imports.
+    - Updated `ImageLoaderSettings`, `ImportFullListScreen`, `GlobalSearchFragment`, `MainActivity`, `OtakuCustomListScreen`, `ImportListScreen`, `DetailsViewModel`, `Pages`, `ReaderComposeImage`, `ReaderSettings`, `DetailsHeader`, and `ViewVideosFragment` to use the new `ComposableUtils`.
