+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/components/CoverCard.kt
@@ -23,8 +23,8 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.programmersbox.kmpuiviews.utils.ComponentState
+import com.programmersbox.kmpuiviews.utils.ComposableUtils
 import com.programmersbox.uiviews.presentation.components.imageloaders.ImageLoaderChoice
-import com.programmersbox.uiviews.utils.ComposableUtils
 import com.programmersbox.uiviews.utils.bounceClick
 import com.programmersbox.uiviews.utils.combineClickableWithIndication
 import com.programmersbox.uiviews.utils.sharedelements.OtakuImageElement
